Morning commit for trip to Washington.
This is horrible practice, but who really cares at this point. I have to head
to Washington and I'm really only using Git right now for myself, so the concept
of best practices is a little out the window.

If you are a Udacity grader, I promise I am much more studious with serious
projects, and this project has already served as a stepping stone for much better
management of repos, resources, etc.